Replace deployTypedContract with overloading method
