+++ .eslintrc.json
@@ -310,12 +310,7 @@
     "no-path-concat": "error",
     "no-proto": "error",
     "no-prototype-builtins": "error",
-    "no-redeclare": [
-      "error",
-      {
-        "builtinGlobals": false
-      }
-    ],
+    "no-redeclare": "warn",
     "no-regex-spaces": "error",
     "no-return-assign": [
       "error",
