+++ waffle-cli/src/index.ts
@@ -3,4 +3,3 @@ export {waffleChai as solidity} from '@ethereum-waffle/chai';
 export {link} from '@ethereum-waffle/compiler';
 export * from '@ethereum-waffle/mock-contract';
 export * from './deployContract';
-export * from './deployTypedContraсt';
