+++ waffle-cli/test/deployTypedContract.ts
@@ -1,14 +1,13 @@
 import {parseEther} from 'ethers/lib/utils';
 import {BasicTokenMock, BasicTokenMockFactory} from '../build';
-import {MockProvider} from '../src';
-import {deployTypedContract} from '../src/deployTypedContraсt';
+import {deployContract, MockProvider} from '../src';
 import {expect} from 'chai';
 
 describe('INTEGRATION: deployTypedContract', () => {
   const [wallet] = new MockProvider().getWallets();
 
   it('successfully deploys contract', async () => {
-    const contract: BasicTokenMock = await deployTypedContract(wallet,
+    const contract: BasicTokenMock = await deployContract(wallet,
       BasicTokenMockFactory,
       [wallet.address, parseEther('100')]
     );
